Refactor LC PoC to prepare copy to FastArduino.
