+++ examples/experiments/Lifecycle/lifecycle.cpp
@@ -1,0 +1,37 @@
+//   Copyright 2016-2020 Jean-Francois Poilpret
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+
+#include "lifecycle.h"
+
+namespace lifecycle
+{
+	AbstractLifeCycle::AbstractLifeCycle(AbstractLifeCycle&& that)
+	{
+		if (that.id_)
+			that.manager()->move_(that.id_, *this);
+	}
+	AbstractLifeCycle::~AbstractLifeCycle()
+	{
+		if (id_)
+			manager()->unregister_(id_);
+	}
+	AbstractLifeCycle& AbstractLifeCycle::operator=(AbstractLifeCycle&& that)
+	{
+		if (id_)
+			manager()->unregister_(id_);
+		if (that.id_)
+			that.manager()->move_(that.id_, *this);
+		return *this;
+	}
+}
